Refactored web server
